EPMLABSBRN-680 BE: fix role restrictions
