+++ src/main/kotlin/com/epam/brn/controller/AdminController.kt
@@ -3,21 +3,28 @@ package com.epam.brn.controller
 import com.epam.brn.dto.BaseResponseDto
 import com.epam.brn.service.StudyHistoryService
 import com.epam.brn.service.UserAccountService
+import com.epam.brn.upload.CsvUploadService
 import io.swagger.annotations.Api
 import io.swagger.annotations.ApiOperation
-import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.format.annotation.DateTimeFormat
+import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
+import org.springframework.web.multipart.MultipartFile
 import java.time.LocalDate
 
 @RestController
 @RequestMapping("/admin")
 @Api(value = "/admin", description = "Contains actions for admin")
-class AdminController(@Autowired val studyHistoryService: StudyHistoryService, @Autowired val userAccountService: UserAccountService) {
+class AdminController(
+    val studyHistoryService: StudyHistoryService,
+    val userAccountService: UserAccountService,
+    private val csvUploadService: CsvUploadService
+) {
 
     @GetMapping("/users")
     @ApiOperation("Get users")
@@ -41,4 +48,13 @@ class AdminController(@Autowired val studyHistoryService: StudyHistoryService, @
         @RequestParam("year", required = true) year: Int
     ) = ResponseEntity.ok()
         .body(BaseResponseDto(data = studyHistoryService.getMonthHistories(userId, month, year)))
+
+    @PostMapping("/loadTasksFile")
+    fun loadExercises(
+        @RequestParam(value = "seriesId") seriesId: Long,
+        @RequestParam(value = "taskFile") file: MultipartFile
+    ): ResponseEntity<BaseResponseDto> {
+        csvUploadService.loadExercises(seriesId, file)
+        return ResponseEntity(HttpStatus.CREATED)
+    }
 }
